Cleaned up/Finish Project 1

+++ Automaton.h
@@ -100,7 +100,7 @@ public:
     }
 
     bool checkEOF(std::string input){
-        return (index >= (int)input.length());
+        return ((unsigned long)index >= input.length());
     }
 };
